requirements.txt
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
asgiref==3.8.1
cachetools==5.5.2
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
cloudinary==1.44.0
Django==5.1.7
django-filter==25.1
djangorestframework==3.16.0
Faker==37.3.0
faker-commerce==1.0.4
google-auth==2.38.0
idna==3.10
packaging==24.2
psycopg2-binary==2.9.10
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
PyJWT==2.10.1
python-decouple==3.8
redis==6.2.0
requests==2.32.3
rsa==4.9
six==1.17.0
sqlparse==0.5.3
stripe==12.0.0
typing_extensions==4.13.2
tzdata==2025.2
urllib3==2.3.0
